shareplate_backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
django-cors-headers==4.9.0
et_xmlfile==2.0.0
geographiclib==2.1
geopy==2.4.1
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary
django-leaflet
